create table event & card, set model

+++ packages/server/migrations/20220115083033-add-event.js
@@ -20,7 +20,7 @@ module.exports = {
           defaultValue: '',
         },
         rent_price: {
-          type: Sequelize.INTEGER,
+          type: Sequelize.DECIMAL,
           defaultValue: 0,
           allowNull: false,
         },
